Add runtime-tunable parameters to filter, distortion, and dynamics opcodes
Extends 8 opcodes with optional runtime parameters for fine-tuning behavior:

Filters:
- FILTER_MOOG: Add max_resonance (self-oscillation threshold) and input_scale (preamp drive)
- FILTER_SALLENKEY: Add clip_threshold (feedback clipping point)

Reverbs:
- REVERB_FREEVERB: Add room_scale (density factor) and room_offset (decay baseline)
- REVERB_DATTORRO: Add input_diffusion and decay_diffusion (smoothing controls)

+++ akkado/include/akkado/codegen.hpp
@@ -14,7 +14,7 @@ namespace akkado {
 
 /// State initialization data for SEQ_STEP and TIMELINE opcodes
 struct StateInitData {
-    std::uint16_t state_id;  // Must match Instruction::state_id (16-bit)
+    std::uint32_t state_id;  // Must match Instruction::state_id (32-bit FNV-1a hash)
     enum class Type : std::uint8_t {
         SeqStep,   // Initialize SeqStepState with timed events
         Timeline   // Initialize TimelineState with breakpoints
